studio/requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
trustcall
langgraph-cli[inmem]
matplotlib
numpy
networkx
seaborn
reportlab
altair
pandasql